cpu/nrf5x: unified rtt driver
